#256 Add sort parameter in the search view
