+++ readthedocs/search/faceted_search.py
@@ -1,22 +1,22 @@
 import logging
 
+from django.conf import settings
 from elasticsearch import Elasticsearch
 from elasticsearch_dsl import FacetedSearch, TermsFacet
-from elasticsearch_dsl.faceted_search import NestedFacet
-from elasticsearch_dsl.query import Bool, SimpleQueryString, Nested, Match
-
-from django.conf import settings
+from elasticsearch_dsl.faceted_search import FacetedResponse, NestedFacet
+from elasticsearch_dsl.query import Bool, Nested, SimpleQueryString
+from elasticsearch_dsl.search import Search
 
 from readthedocs.core.utils.extend import SettingsOverrideObject
-from readthedocs.search.documents import (
-    PageDocument,
-    ProjectDocument,
-)
-
+from readthedocs.projects.constants import PRIVATE
+from readthedocs.search.documents import PageDocument, ProjectDocument
 
 log = logging.getLogger(__name__)
 
-ALL_FACETS = ['project', 'version', 'role_name', 'language', 'index']
+ALL_FACETS = [
+    'project', 'version', 'role_name', 'language', 'index', 'publisher',
+    'publisher_project',
+]
 
 
 class RTDFacetedSearch(FacetedSearch):
@@ -70,7 +70,7 @@ class RTDFacetedSearch(FacetedSearch):
 
         for operator in self.operators:
             query_string = SimpleQueryString(
-                query=query, fields=self.fields, default_operator=operator
+                query=query, fields=self.fields, default_operator=operator,
             )
             all_queries.append(query_string)
 
@@ -82,7 +82,11 @@ class RTDFacetedSearch(FacetedSearch):
 
 
 class ProjectSearchBase(RTDFacetedSearch):
-    facets = {'language': TermsFacet(field='language')}
+    facets = {
+        'language': TermsFacet(field='language'),
+        'publisher': TermsFacet(field='publisher'),
+        'publisher_project': TermsFacet(field='publisher_project'),
+    }
     doc_types = [ProjectDocument]
     index = ProjectDocument._doc_type.index
     fields = ('name^10', 'slug^5', 'description')
@@ -97,6 +101,8 @@ class PageSearchBase(RTDFacetedSearch):
             'domains',
             TermsFacet(field='domains.role_name')
         ),
+        'publisher': TermsFacet(field='publisher'),
+        'publisher_project': TermsFacet(field='publisher_project'),
     }
     doc_types = [PageDocument]
     index = PageDocument._doc_type.index
@@ -203,6 +209,14 @@ class PageSearchBase(RTDFacetedSearch):
             query=bool_query
         )
         return nested_query
+
+    def search(self):
+        """Construct the Search object and return a faceted search response."""
+        s = Search(
+            doc_type=self.doc_types, index=self.index, using=self.using,
+            extra={'min_score': getattr(settings, 'ES_SEARCH_FILE_MIN_SCORE', 1)},
+        ).exclude("term", privacy_level=PRIVATE)
+        return s.response_class(FacetedResponse)
 
 
 class PageSearch(SettingsOverrideObject):
