+++ readthedocs/urls.py
@@ -11,13 +11,12 @@ from django.views.generic.base import RedirectView, TemplateView
 
 from readthedocs.core.urls import core_urls, deprecated_urls, docs_urls
 from readthedocs.core.views import (
-    HomepageView,
     SupportView,
     do_not_track,
     server_error_404,
     server_error_500,
 )
-from readthedocs.docsitalia.search import views as search_views
+from readthedocs.search import views as search_views
 from readthedocs.search.api import PageSearchAPIView
